feat:add hybrid_search to rag demo (#46)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->
Add hybrid_search to rag demo
close #45


## Solution Description
<!-- Please clearly and concisely describe your solution. -->
Add a logic which user can choose weather to use hybrid_search or not