feat:update readme of rag demo (#61)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->
The rag demo has supported hybrid search.The readme need to be updated.
close #60


## Solution Description
<!-- Please clearly and concisely describe your solution. -->